+++ matchbox-frontend/src/app/validate/structure-definition.ts
@@ -1,0 +1,17 @@
+export class StructureDefinition {
+
+  // The canonical
+  canonical: string;
+
+  // The title
+  title: string;
+
+  // The ImplementationGuide identifier
+  igId: string;
+
+  // The ImplementationGuide version
+  igVersion: string;
+
+  // Whether the ImplementationGuide version is the current one (the higher version)
+  isCurrent: boolean;
+}